# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Discutex API",
        "description": "API for managing threads, responses, and votes",
        "termsOfService": "https://www.example.com/terms",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/responses/thread/{thread_id}": {
            "get": {
                "tags": [
                    "Responses"
                ],
                "summary": "Get responses for a specific thread",
                "description": "Retrieves all responses for a thread using its ID.",
                "operationId": "getResponsesByThread",
                "parameters": [
                    {
                        "name": "thread_id",
                        "in": "path",
                        "description": "The thread ID to fetch associated responses.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Responses found successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "responses": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "content": {
                                                        "type": "string"
                                                    },
                                                    "user_id": {
                                                        "type": "integer"
                                                    },
                                                    "thread_id": {
                                                        "type": "integer"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Thread not found."
                    },
                    "500": {
                        "description": "Error while fetching responses."
                    }
                }
            }
        },
        "/responses/{thread_id}": {
            "get": {
                "tags": [
                    "Responses"
                ],
                "summary": "Get responses for a thread (with user data)",
                "description": "Retrieves all responses for a specific thread, including the user data for each response.",
                "operationId": "getResponsesWithUser",
                "parameters": [
                    {
                        "name": "thread_id",
                        "in": "path",
                        "description": "The thread ID to fetch responses.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Responses found successfully with user data.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "responses": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "content": {
                                                        "type": "string"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "email": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "thread_id": {
                                                        "type": "integer"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    },
                                                    "updated_at": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error while fetching responses."
                    }
                }
            },
            "post": {
                "tags": [
                    "Responses"
                ],
                "summary": "Create a new response for a thread",
                "description": "Create a new response for a specific thread.",
                "operationId": "createResponse",
                "parameters": [
                    {
                        "name": "thread_id",
                        "in": "path",
                        "description": "The thread ID to create a response in.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "description": "Content of the response",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Response created successfully."
                    },
                    "400": {
                        "description": "Invalid request data."
                    },
                    "500": {
                        "description": "Error while creating response."
                    }
                }
            }
        },
        "/responses/user/{response_id}": {
            "get": {
                "tags": [
                    "Responses"
                ],
                "summary": "Get the user who made a specific response",
                "description": "Retrieves the user related to a response using the response ID.",
                "operationId": "getUserByResponse",
                "parameters": [
                    {
                        "name": "response_id",
                        "in": "path",
                        "description": "The response ID to get the associated user data.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Response or user not found."
                    },
                    "500": {
                        "description": "Error while fetching user data."
                    }
                }
            }
        },
...
    },
...
}